% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowlayers2-package.R
\docType{package}
\name{flowlayers2-package}
\alias{flowlayers2}
\alias{flowlayers2-package}
\title{flowlayers2: ATFutures Development Package}
\description{
WHO metapackage for generating layers of pedestrian flows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ATFutures-labs/flowlayers2}
  \item Report bugs at \url{https://github.com/ATFutures-labs/flowlayers2/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\keyword{internal}
