% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowlayers.R
\name{get_bp}
\alias{get_bp}
\title{get_bp}
\usage{
get_bp(city)
}
\arguments{
\item{city}{Name of city for which to obtain bounding polygon}
}
\value{
Matrix of coordinates of bounding polygon
}
\description{
get bounding polygons for cities (Accra, Kathmandu)
}
