% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowlayers.R
\name{network_flow}
\alias{network_flow}
\title{network_flow}
\usage{
network_flow(city, save = TRUE, quiet = FALSE)
}
\arguments{
\item{city}{Name of city for which to obtain bounding polygon}

\item{save}{If `TRUE`, save data to internal `who-data` directory.}

\item{quiet}{If `FALSE`, display progress information on screen.}
}
\description{
Generate pedestrian flows for one city (Accra, Kathmandu).
}
