% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowlayers.R
\name{osm_buildings}
\alias{osm_buildings}
\title{osm_buildings}
\usage{
osm_buildings(city, save = TRUE, quiet = FALSE)
}
\arguments{
\item{city}{Name of city for which to obtain bounding polygon}

\item{save}{If `TRUE`, save data to internal `who-data` directory.}

\item{quiet}{If `FALSE`, display progress information on screen.}
}
\description{
osm_buildings
}
