% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowlayers.R
\name{osm_bus_centrality}
\alias{osm_bus_centrality}
\title{osm_bus_centrality}
\usage{
osm_bus_centrality(city)
}
\arguments{
\item{city}{Name of city for which to obtain bounding polygon}
}
\description{
Calculate centrality of bus networks for cities, and assign vertex
centrality values to each bus stop, in order to use these as proxies for
numbers of passengers.
}
