% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowlayers.R
\name{osm_bus_network}
\alias{osm_bus_network}
\title{osm_bus_network}
\usage{
osm_bus_network(city)
}
\arguments{
\item{city}{Name of city for which to obtain bounding polygon}
}
\description{
Get bus network data for cities (Accra, Kathmandu).
}
