% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowlayers.R
\name{osm_bus_stops}
\alias{osm_bus_stops}
\title{osm_bus_stops}
\usage{
osm_bus_stops(city, save = TRUE)
}
\arguments{
\item{city}{Name of city for which to obtain bounding polygon}

\item{save}{If `TRUE`, save data to internal `who-data` directory.}
}
\description{
Get bus stop data for cities (Accra, Kathmandu).
}
