% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowlayers.R
\name{osm_street_network}
\alias{osm_street_network}
\title{Get street network for WHO# cities (Accra and Kathmandu)}
\usage{
osm_street_network(city, save = TRUE, quiet = FALSE)
}
\arguments{
\item{city}{Name of city for which to obtain bounding polygon}

\item{save}{If `TRUE`, save data to internal `who-data` directory.}

\item{quiet}{If `FALSE`, display progress information on screen.}
}
\value{
SC-formatted street network for nominated city.
}
\description{
Get street network for WHO# cities (Accra and Kathmandu)
}
